requirements.txt
opencv-python
numpy
ultralytics
python-dotenv
opencv-python-headless 